Fixed empty finalAction array issue
* Updated Surfile content for supporting FinalAction.
* Updated surmagic binary
