+++ Demo/SM/Surfile
@@ -1,29 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-	<dict>
-		<key>output_path</key>
-		<string>build</string>
-		<key>framework</key>
-		<string>xcf</string>
-		<key>targets</key>
-		<array>
-			<dict>
-				<key>sdk</key>
-				<string>iOS</string>
-				<key>workspace</key>
-				<string>surmagic-demo.xcworkspace</string>
-				<key>scheme</key>
-				<string>xcf</string>
-			</dict>
-			<dict>
-				<key>sdk</key>
-				<string>iOSSimulator</string>
-				<key>workspace</key>
-				<string>surmagic-demo.xcworkspace</string>
-				<key>scheme</key>
-				<string>xcf</string>
-			</dict>
-		</array>
-	</dict>
+    <dict>
+        <key>output_path</key>
+        <string>build</string>
+        <key>framework</key>
+        <string>xcf</string>
+        <key>targets</key>
+        <array>
+            <dict>
+                <key>sdk</key>
+                <string>iOS</string>
+                <key>workspace</key>
+                <string>surmagic-demo.xcworkspace</string>
+                <key>scheme</key>
+                <string>xcf</string>
+            </dict>
+            <dict>
+                <key>sdk</key>
+                <string>iOSSimulator</string>
+                <key>workspace</key>
+                <string>surmagic-demo.xcworkspace</string>
+                <key>scheme</key>
+                <string>xcf</string>
+            </dict>
+        </array>
+        <key>finalActions</key>
+        <array>
+            <string>openDirectory</string>
+        </array>
+    </dict>
 </plist>
